refac the GaEvent
